fix(statsPage): use Preload for assessment statistics page
